make command log size configurable
